refactor: #8 내보내기 방식 수정
