+++ src/app/layout.tsx
@@ -1,20 +1,17 @@
-// src/app/layout.tsx
+// src/app/layout.tsx (서버 컴포넌트)
+import { ThemeProvider } from '@/modules/theme/ThemeProvider';
 
-import { ThemeProvider } from 'next-themes';
 import '../styles/globals.css';
-import { Header } from '@/components/Layout/Header';
-import { Footer } from '@/components/Layout/Footer';
+import { Header, Footer } from '@/components';
 
-export const RootLayout = ({ children }: { children: React.ReactNode }) => {
+const RootLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <html lang="ko">
-      <head></head>
       <body>
-        <ThemeProvider attribute="class">
+      <ThemeProvider>
           <div className="flex min-h-screen flex-col">
             <Header />
             <div className="flex-1">{children}</div>
-
             <Footer />
           </div>
         </ThemeProvider>
@@ -22,3 +19,5 @@ export const RootLayout = ({ children }: { children: React.ReactNode }) => {
     </html>
   );
 };
+
+export default RootLayout;
