+++ src/app/page.tsx
@@ -1,5 +1,7 @@
 const containerClass = 'flex flex-col items-center justify-center space-y-8 min-h-full';
 
-export const Home = () => {
+const Home = () => {
   return <div className={containerClass}>메인페이지</div>;
 };
+
+export default Home;
