+++ src/components/index.ts
@@ -1,1 +1,4 @@
 export { Button } from './Button/Button';
+
+export { default as Header } from './Layout/Header';
+export { default as Footer } from './Layout/Footer';
