+++ src/modules/theme/ThemeProvider.tsx
@@ -1,0 +1,11 @@
+'use client';
+ 
+import { ThemeProvider as NextThemesProvider, ThemeProviderProps } from 'next-themes';
+ 
+export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
+  return (
+    <NextThemesProvider attribute='class' defaultTheme='system' {...props}>
+      {children}
+    </NextThemesProvider>
+  );
+}
